Refactor a base class
