+++ packages/ai/src/api.ts
@@ -45,6 +45,7 @@ import { logger } from './logger';
 
 export { ChatSession } from './methods/chat-session';
 export { LiveSession } from './methods/live-session';
+export { TemplateChatSession } from './methods/template-chat-session';
 export * from './requests/schema-builder';
 export { ImagenImageFormat } from './requests/imagen-image-format';
 export {
