+++ packages/ai/src/public-types.ts
@@ -19,7 +19,6 @@ import { FirebaseApp } from '@firebase/app';
 import { Backend } from './backend';
 
 export * from './types';
-export { TemplateChatSession } from './methods/template-chat-session';
 
 /**
  * An instance of the Firebase AI SDK.
